wip lab2

+++ Lab1_427/Core/Src/button_helper.c
@@ -12,19 +12,16 @@ int is_button_pressed() {
 	static int prev_state = 0;
 	static int last_release = 0;
 	int current_tick = HAL_GetTick();
-	int current_state = HAL_GPIO_ReadPin(Button_GPIO_Port, Button_Pin) == GPIO_PIN_RESET ?
-			1 : 0;
+	uint8_t current_state =
+			HAL_GPIO_ReadPin(Button_GPIO_Port, Button_Pin) == GPIO_PIN_RESET ?
+					1 : 0;
 
-	if(prev_state && current_state)
-	{
-		if(current_tick - last_release > 50)
-		{
+	if (prev_state && current_state) {
+		if (current_tick - last_release > 50) {
 			return 1;
 		}
 		return 0;
-	}
-	else
-	{
+	} else {
 		prev_state = current_state;
 		last_release = current_tick;
 		return 0;
